# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "description": "Crea un nuevo usuario con los datos proporcionados.",
                "operationId": "c0eaa9600f14ac767e20818010954c59",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre del usuario"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "contraseña"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario creado exitosamente"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "Token de acceso generado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Campos inválidos o faltantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Existen campos vacíos"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Iniciar sesión de usuario",
                "description": "Inicia sesión de usuario con las credenciales proporcionadas.",
                "operationId": "e45287a3dc02f4633476dbf218bb9ef7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "contraseña"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inicio de sesión exitoso"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "Token de acceso generado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas o faltantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Correo electrónico y contraseña no coinciden con nuestros registros"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/candidatos": {
            "get": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Listado de candidatos con información adicional",
                "description": "Muestra un listado de los candidatos con detalles sobre la lista y el tipo de candidato.",
                "operationId": "13297296e2978e75b663c6571e88f2da",
                "responses": {
                    "200": {
                        "description": "Listado de candidatos con información adicional",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "candidatos": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Descripción del candidato"
                                                    },
                                                    "idlista": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "lista": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "descripcion": {
                                                                "type": "string",
                                                                "example": "Descripción de la lista"
                                                            },
                                                            "numero": {
                                                                "type": "string",
                                                                "example": "Número de la lista"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "idtipocandidato": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "tipo_candidato": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "descripcion": {
                                                                "type": "string",
                                                                "example": "Descripción del tipo de candidato"
                                                            },
                                                            "icon": {
                                                                "type": "string",
                                                                "example": "icono-del-tipo"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/lista-candidatos-total-votos": {
            "get": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Listado de candidatos con total de votos por lista",
                "description": "Muestra un listado de los candidatos con la suma total de votos por lista.",
                "operationId": "dc274efc8a558e537bfb929d136ea3b3",
                "responses": {
                    "200": {
                        "description": "Listado de candidatos con total de votos por lista",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "lista_candidatos": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "idlista": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre_candidato": {
                                                        "type": "string",
                                                        "example": "Nombre del candidato"
                                                    },
                                                    "total_votos": {
                                                        "type": "integer",
                                                        "example": 100
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}